Committing my solution to the first ESHS summer robotics exercise.
The problem statement is:

"Write a program that asks the user for the number of dice to roll, then have
it roll that many dice and output the sum."

I did this using idiomatic modern C++ constructs:

- I generate cryptographically secure random numbers using the C++11 <random> header.
- The program is robust (as far as I can tell) to bad input.
- I used some of C++'s functional (well, functor-related) algorithms:
  * std::accumulate to generate the sum,
  * std::generate to populate the rolls array, and
  * std::copy to copy the integers, one at a time, into an ostream_iterator in
    order to print the addition formula at the end.
- Note the use of bind() to turn a unary function ("distribution(generator)")
  into a nullary function ("rollOneDie()").

If the program has any bug at all, it's that it dynamically allocates memory
on the heap in order to store the dice rolls in a container so I can use the
aforementioned fancy functions on it.  So the user can quite easily crash the
program by giving it a stupidly large number of dice to roll.  The naive
looping implementation does not have that problem.

Also added a gitignore because seeing the backup file in the git status
annoyed me.