Adding VSCode settings for C++.
I installed the C/C++ Intellisense extension for VSCode, then configured
a handful of build tasks -- one to compile, one to compile and deploy,
and one to debug.  Use CTRL+SHIFT+P -> Tasks: Run Build Task.

But I'm still getting green wriggly lines for the "include main.h"
line, so clearly the changes I made to the include path are not being
respected.  The code still builds, and I'm not a fan of VSCode, so
this seems good enough for now.