Remove the aggressive PROS gitignores.
While it was nice to only have the user-editable files in the project, I
miscalculated: PROSv5 currently offers no way to take an incomplete project
and "bootstrap" it into a compiled, working PROS project.  (If it were
designed properly, the command would be "prosv5 conduct new . --force".)

So for now, it's better to version the PROS files that "prosv5 conduct new"
generates.  However, I'm still leaving out the files generated by "prosv5
make" -- in short, I'm versioning the libraries, not the executables.