Committing files which I mistakenly left out of the build.
I used to think that all you needed for a ProsV5 project were the human-
edited files.  My reasoning was that I could simply tell Pros "generate
a standard project in this existing folder" and then overwrite the five
main files with the existing ones.

But ProsV5 does not work that way.  It really, _really_ only wants to
create the stub project once, so I can't just ignore the stuff it
generates one time during "prosv5 conduct new ." -- it has no ability
to recreate what's missing.

Thus I changed the .gitignore and added the rest of what we need.