The entrance/exit hunting function seems to work!
One of the keys to getting it to function was eliminating the return values,
storing all the parameters that were needed after recursion in one struct, and
passing that by reference.  All the alteration to the data structure only
occurs at the end, after all neighbors are processed.

The algorithm isn't very efficient, but it really does find the optimal
entrance and exit for the maze.  It uses a special =drawing routine to cut
holes in the border without damaging adjacent walls.

The function returns the (unit) coordinates of the entrance and exit it just
cut; I suppose that would be useful for labels.