Replace GenerateOverExisting() with NewMazeOverExisting().
The idea is that generation has as its sole input the m.cells array, and
whatever's already there is taken into account appropriately.  We do not care
what it is!

We'll still be using odd-numbered unit rectangles as always.  But which ones
are legal to draw walls from, and which are off-limits?  I'm still trying to
figure that out right now.