Add _support_ (just the support) for multiple thickness values.  They must be unique.
The thickness values will cause the maze to be generated with successively
smaller and smaller walls, each using the empty space from the previous
iteration of the maze.

At least in theory.

Actually passing in more than one thickness right now hangs the program; I'll
work on fixing that next.