Committing a program that makes a 3D rotating cube...in ASCII.
It was quite a bit of work getting this to render, starting with teaching the
class how to compile ncurses from scratch.  From there, we set a single red
"pixel" at (5, 5) (as I did back when I was first learning 3D graphics); a
projection formula and a rudimentary line-drawing algorithm later, and here we
are: a full-blown wireframe cube rotating according to arbitrary Euler angles
passed in as command-line arguments.

The program has flaws.  It's not universal; it depends on and links to a local
copy of the ncurses library instead of detecting the system version like it's
supposed to.  But it is a milestone nonetheless, and worthy of committing.